+++ generators/app/templates/src/js/app.js
@@ -1,5 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import Landing from './components/landing';
-
-ReactDOM.render(<Landing />, document.getElementById('app'));