+++ generators/app/templates/test/components/landing.js
@@ -1,9 +1,0 @@
-import Landing from '../../src/js/components/landing';
-
-describe('landing.js', () => {
-  describe('<Landing />', () => {
-    it('has no tests', () => {
-      // See Enzyme documentation: https://github.com/airbnb/enzyme
-    });
-  });
-});